Replace Link components (#433)
