+++ .changeset/cold-spoons-taste.md
@@ -1,0 +1,5 @@
+---
+'@utilitywarehouse/ds-codemod': minor
+---
+
+Add codemod to update lab imports
