+++ .changeset/popular-pillows-appear.md
@@ -1,0 +1,5 @@
+---
+'@utilitywarehouse/web-ui': major
+---
+
+Replace `Link` and `TextLink` components with lab components
