+++ .changeset/selfish-pears-prove.md
@@ -1,0 +1,5 @@
+---
+'@utilitywarehouse/web-ui': minor
+---
+
+Implement `textTransform` for `Link` & `TextLink` components
