+++ .changeset/twenty-pans-crash.md
@@ -1,0 +1,5 @@
+---
+'@utilitywarehouse/web-ui': patch
+---
+
+Remove `justify-content: center` from link components.
