+++ packages/codemod/transforms/web-ui/v1/__testfixtures__/update-lab-imports.input.tsx
@@ -1,0 +1,14 @@
+// this is a comment at the beginning of the file
+import * as React from 'react';
+import { Link, TextLink } from '@utilitywarehouse/web-ui/dist/lab';
+
+const Component = () => (
+  <div>
+    <Link>Hamburgefons</Link>
+    <TextLink>Hamburgefons</TextLink>
+    <Link size='small'>Hamburgefons</Link>
+    <Link size='large'>Hamburgefons</Link>
+  </div>
+);
+
+export default Component;
