+++ packages/codemod/transforms/web-ui/v1/__tests__/update-lab-imports.test.js
@@ -1,0 +1,10 @@
+/* eslint-disable no-undef */
+const defineTest = require('jscodeshift/dist/testUtils').defineTest;
+
+const NAME = 'update-lab-imports';
+
+describe(NAME, () => {
+  defineTest(__dirname, NAME, {}, undefined, {
+    parser: 'tsx',
+  });
+});
