+++ packages/web-ui/src/Alert/AlertLink.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 import { colors } from '@utilitywarehouse/colour-system';
-import { TextLink } from '../lab/TextLink';
-import type { TextLinkProps } from '../lab/TextLink';
+import { TextLink } from '../TextLink';
+import type { TextLinkProps } from '../TextLink';
 import { styled } from '../theme';
 import { fontWeights } from '../tokens';
 import { colorSchemeParentSelector, px, withGlobalPrefix } from '../utils';
