+++ packages/web-ui/src/Link/Link.props.ts
@@ -1,5 +1,6 @@
 import { ComponentPropsWithoutRef } from 'react';
-import { Responsive } from '../../types';
+import { Responsive } from '../types';
+import { TextProps } from '../Text';
 
 export interface LinkProps extends ComponentPropsWithoutRef<'a'> {
   /**
@@ -11,4 +12,5 @@ export interface LinkProps extends ComponentPropsWithoutRef<'a'> {
    * Change the default rendered element for the one passed as a child, merging their props and behavior.
    */
   asChild?: boolean;
+  textTransform?: TextProps['textTransform'];
 }
