+++ packages/web-ui/src/Link/index.ts
@@ -1,2 +1,2 @@
 export { Link } from './Link';
-export type { LinkProps } from './Link';
+export type { LinkProps } from './Link.props';
