+++ packages/web-ui/src/TextLink/index.ts
@@ -1,2 +1,2 @@
 export { TextLink } from './TextLink';
-export type { TextLinkProps } from './TextLink';
+export type { TextLinkProps } from './TextLink.props';
