+++ packages/web-ui/src/lab/Button/Button.docs.mdx
@@ -24,7 +24,7 @@ import { DocsHeader } from '../../storybook-utils';
 - [Usage with Next.js](#usage-with-nextjs)
   - [Next.js v13](#nextjs-v13)
   - [Next.js before v13](#nextjs-before-v13)
-- [API Reference](#api-reference)
+- [Props](#props)
 
 ## Usage
 
@@ -326,6 +326,6 @@ import { Button } from '@utilitywarehouse/web-ui/dist/lab';
 </NextLink>
 ```
 
-## API Reference
+## Props
 
 <ArgTypes of={Stories} />
