--- packages/web-ui/src/lab/Link/Link.docs.mdx
@@ -1,187 +1,0 @@
-import { Meta, Canvas, ArgTypes } from '@storybook/blocks';
-import * as Stories from './Link.stories';
-import { DocsHeader } from '../../storybook-utils';
-
-<Meta of={Stories} />
-
-<DocsHeader
-  lab
-  componentName="Link"
-  stories={Stories}
-  figmaLink="https://www.figma.com/design/4FFYTLWJ2hQpj36JplQQUw/UW-Web-UI?node-id=6834-38146"
-/>
-
-<Canvas of={Stories.SimpleExample} sourceState="none" />
-
-- [Usage](#usage)
-- [Size](#size)
-- [With icons](#with-icons)
-- [Accessibility](#accessibility)
-- [Semantic HTML](#semantic-html)
-- [Full Width Link](#full-width-link)
-- [Usage with Next.js](#usage-with-nextjs)
-  - [Next.js v13](#nextjs-v13)
-  - [Next.js before v13](#nextjs-before-v13)
-- [API Reference](#api-reference)
-
-## Usage
-
-This `Link` component is part of the lab submodule, a way for us to make new
-components available prior to an official major release. This helps us bed in
-new components before introducing breaking changes, and to avoid blocking the
-availability of these components for designers.
-
-```tsx
-import { Box } from '@utilitywarehouse/web-ui';
-import { Link } from '@utilitywarehouse/web-ui/dist/lab';
-
-[...]
-
-<Box>
-  <Link>Link</Link>
-</Box>
-```
-
-## Size
-
-The size prop controls the size of the `Link`. This is a responsive prop and
-can be used to display different sizes at different breakpoints.
-
-<Canvas of={Stories.ResponsiveSize} />
-
-```tsx
-<Link
-  size={{
-    mobile: 'small',
-    desktop: 'large',
-  }}
-/>
-```
-
-## With icons
-
-You can nest icons directly inside the `Link`. An appropriate gap is provided
-automatically, and the icon will inherit the appropriate colours.
-
-This button is intended to be used with the [UW Icons](https://uw-icons.vercel.app/) packages.
-Small links should use 16px icons (small), while large links should use 24px
-icons (medium).
-
-```tsx
-<Link size="large">
-  <SettingsMediumIcon />
-  Edit account
-</Link>
-```
-
-<Canvas of={Stories.WithIcons} />
-
-## Accessibility
-
-Icons from the UW icons packages have `aria-hidden="true"` set by default. If
-you are using an icon from outside these packages, and for purely decorative
-reasons, please ensure it has this attribute set so that the icon is hidden
-from screen readers.
-
-## Semantic HTML
-
-> If it goes somewhere it's a link, if it does something it's a button.
-
-A semantic HTML `a` is rendered by default, however you can change the
-underlying HTML element by using the `asChild` prop.
-
-When `asChild` is set to true, we will not render a default DOM element,
-instead cloning the child and passing it the props and behaviour required to
-make it functional.
-
-Read more about this idea in the [Radix UI composition docs](https://www.radix-ui.com/primitives/docs/guides/composition).
-
-```tsx
-// You may need to disable this eslint warning:
-// eslint-disable-next-line jsx-a11y/anchor-is-valid
-<Link asChild>
-  <button type="button" onClick={onClick}>
-    View UW services
-    <OpenMediumIcon />
-  </button>
-</Link>
-```
-
-<Canvas of={Stories.AsButton} />
-
-## Full Width Link
-
-Use layout components, like `Flex`, to define your UI layouts.
-
-<Canvas of={Stories.FullWidth} />
-
-```tsx
-<Flex direction="column" align="stretch">
-  <Link href="#">Link</Link>
-</Flex>
-```
-
-## Usage with Next.js
-
-### Next.js v13
-
-The Next.js `Link` component behaviour has changed in v13, so that an `<a>` is
-no longer required as a child. You can render the Web UI `Link` component as a
-Next.js `Link` component using `asChild`:
-
-```tsx
-import NextLink from 'next/link';
-import { Link } from '@utilitywarehouse/web-ui/dist/lab';
-
-[...]
-
-{/* // eslint-disable-next-line jsx-a11y/anchor-is-valid */}
-<Link asChild>
-  <NextLink href={href} onClick={onClick}>
-    {title}
-  </NextLink>
-</Link>
-```
-
-You can also use the `legacyBehavior` prop directly on the Next.js Link component:
-
-```tsx
-import NextLink from 'next/link';
-import { Link } from '@utilitywarehouse/web-ui/dist/lab';
-
-[...]
-
-<NextLink href={href} passHref onClick={onClick} legacyBehavior>
-  {/* // eslint-disable-next-line jsx-a11y/anchor-is-valid */}
-  <Link>{title}</Link>
-</NextLink>
-```
-
-And if you want to set this behavior globally you can use the following Next.js
-configuration:
-
-```
-{
-  experimental: {
-    newNextLinkBehavior: false
-  }
-}
-```
-
-### Next.js before v13
-
-```tsx
-import NextLink from 'next/link';
-import { Link } from '@utilitywarehouse/web-ui/dist/lab';
-
-[...]
-
-<NextLink href={href} passHref onClick={onClick}>
-  {/* // eslint-disable-next-line jsx-a11y/anchor-is-valid */}
-  <Link>{title}</Link>
-</NextLink>
-```
-
-## API Reference
-
-<ArgTypes of={Stories} />
