--- packages/web-ui/src/lab/Link/index.ts
@@ -1,2 +1,0 @@
-export { Link } from './Link';
-export type { LinkProps } from './Link.props';
