--- packages/web-ui/src/lab/TextLink/index.ts
@@ -1,2 +1,0 @@
-export { TextLink } from './TextLink';
-export type { TextLinkProps } from './TextLink.props';
