+++ packages/web-ui/src/lab/index.ts
@@ -1,3 +1,1 @@
 export * from './Button';
-export * from './Link';
-export * from './TextLink';
